Merge Request: 20170214完成11个代码的修改及测试
Created By: @lu_zhang
Accepted By: @shniu
URL: https://coding.net/u/digcreditdev/p/featurefactory/git/merge/25

+++ studio/lp_dataocean_handle/lp_airfare_sum_12.py
@@ -27,15 +27,14 @@ class Handle(object):
         """
         try:
             result = {
-                'airfare_sum_12': 9999
+                'airfare_sum_12': 9999.0
             }
 
             if self.data['result'] == '00':
                 average_price = self.data['content'].get('average_price', None)
                 flight_times = self.data['content'].get('flight_times', None)
                 if average_price and flight_times:
-                    result['airfare_sum_12'] = average_price * flight_times
-
+                    result['airfare_sum_12'] = float(average_price * flight_times)
         except Exception as e:
             logging.error(e.message)
